+++ frontend/.storybook/preview.js
@@ -1,5 +1,7 @@
+import { ThemeProvider } from '@emotion/react';
 import { RecoilRoot } from 'recoil';
 import GlobalStyle from '../src/style/GlobalStyle';
+import theme from '../src/style/theme';
 
 export const parameters = {
   actions: { argTypesRegex: '^on[A-Z].*' },
@@ -18,9 +20,11 @@ export const parameters = {
 
 export const decorators = [
   (Story) => (
+    <ThemeProvider theme={theme.light}>
       <GlobalStyle />
       <RecoilRoot>
         <Story />
       </RecoilRoot>
+    </ThemeProvider>
   ),
 ];
